+++ github-projects-md/scripts/config/badgeMap.js
@@ -1,26 +1,28 @@
 export const badgeMap = {
-  next: "![Next JS](https://img.shields.io/badge/Next.js-000000?style=for-the-badge&logo=nextdotjs&logoColor=white)",
-  react:
+  "Next JS":
+    "![Next JS](https://img.shields.io/badge/Next.js-000000?style=for-the-badge&logo=nextdotjs&logoColor=white)",
+  React:
     "![React](https://img.shields.io/badge/React-20232A?style=for-the-badge&logo=react&logoColor=61DAFB)",
-  electron:
+  Electron:
     "![Electron](https://img.shields.io/badge/Electron-2C2E3B?style=for-the-badge&logo=electron&logoColor=9FEAF9)",
-  typescript:
+  TypeScript:
     "![TypeScript](https://img.shields.io/badge/TypeScript-007ACC?style=for-the-badge&logo=typescript&logoColor=white)",
-  javascript:
+  JavaScript:
     "![JavaScript](https://img.shields.io/badge/JavaScript-F7DF1E?style=for-the-badge&logo=javascript&logoColor=000)",
-  tailwind:
+  "Tailwind CSS":
     "![Tailwind CSS](https://img.shields.io/badge/Tailwind%20CSS-38B2AC?style=for-the-badge&logo=tailwindcss&logoColor=white)",
-  styled:
+  "Styled Components":
     "![Styled Components](https://img.shields.io/badge/Styled%20Components-DB7093?style=for-the-badge&logo=styledcomponents&logoColor=white)",
-  css: "![CSS](https://img.shields.io/badge/CSS-1572B6?style=for-the-badge&logo=css&logoColor=white)",
-  scss: "![SCSS](https://img.shields.io/badge/SCSS-CC6699?style=for-the-badge&logo=sass&logoColor=white)",
-  sass: "![Sass](https://img.shields.io/badge/Sass-CC6699?style=for-the-badge&logo=sass&logoColor=white)",
-  zustand:
-    "![Zustand](https://img.shields.io/badge/Zustand-000000?style=for-the-badge&logo=javascript&logoColor=white)",
-  redux:
+  HTML: "![HTML](https://img.shields.io/badge/HTML-E34F26?style=for-the-badge&logo=html5&logoColor=white)",
+  CSS: "![CSS](https://img.shields.io/badge/CSS-1572B6?style=for-the-badge&logo=css3&logoColor=white)",
+  SASS: "![Sass](https://img.shields.io/badge/Sass-CC6699?style=for-the-badge&logo=sass&logoColor=white)",
+  SCSS: "![SCSS](https://img.shields.io/badge/SCSS-CC6699?style=for-the-badge&logo=sass&logoColor=white)",
+  "Redux Toolkit":
     "![Redux Toolkit](https://img.shields.io/badge/Redux%20Toolkit-764ABC?style=for-the-badge&logo=redux&logoColor=white)",
+  Zustand:
+    "![Zustand](https://img.shields.io/badge/Zustand-000000?style=for-the-badge&logo=javascript&logoColor=white)",
   NextIntl:
     "![NextIntl](https://img.shields.io/badge/NextIntl-000000?style=for-the-badge&logo=nextdotjs&logoColor=white)",
-  i18next:
+  i18Next:
     "![i18Next](https://img.shields.io/badge/i18Next-26A69A?style=for-the-badge&logo=javascript&logoColor=white)",
 };